\name{divnet-package}
\alias{divnet-package}
\alias{divnet}
\docType{package}
\title{
Diversity indices for networks
}
\description{
Bla bla bla bla bla blab  bla
}
\details{
divnet is bla bla  bla.
}
\author{
Authors: Stéphane Dray, Vincent Miele, Marc Ohlmann, Wilfried Thuiller
  
Maintainer: Wilfried Thuiller <wilfried.thuiller@univ-grenoble-alpes.fr>
}
\references{

}
\keyword{ package }
\keyword{ ecology }
\keyword{ diversity }
\keyword{ network }
%\seealso{
%Optional links to other man pages, e.g. \code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}
%\examples{
%}
