\name{econetwork-package}
\alias{econetwork-package}
\alias{econetwork}
\docType{package}
\title{
Analyzing Ecological Networks
}
\description{
A collection of advanced tools, methods and models specifically
designed  for  analyzing different  types  of  ecological  networks.
}
\details{
econetwork is designed  for  analyzing different  types  of  ecological
networks - especially  antagonistic  (food  webs,  host-parasite),
mutualistic (plant-pollinator, plant-fungus, etc.)  and competitive
networks, as well as their variability in time and space. Statistical
models are developed to describe and understand the mechanisms that
determine species interactions, and to decipher the organization of
these (multi-layer) ecological networks.
}
\author{
Authors: Stephane Dray, Catherine Matias, Vincent Miele, Marc Ohlmann, Wilfried Thuiller
Maintainer: Vincent Miele <vincent.miele@univ-lyon1.fr>
}
\references{
Marc Ohlmann, Vincent Miele, Stephane Dray, Loic Chalmandrier, Louise O'Connor & Wilfried Thuiller, Diversity indices for ecological networks: a unifying framework using Hill numbers. Ecology Letters (2019) <doi:10.1111/ele.13221>
Ana M. Martin Gonzalez, Diego P. Vazquez, Rodrigo Ramos-Jiliberto, Sang  Hoon Lee & Vincent Miele, Core-periphery structure in mutualistic networks: an epitaph for nestedness? BiorXiv (2020) <doi:10.1101/2020.04.02.021691>
}
\keyword{package}
\keyword{ecology}
\keyword{diversity}
\keyword{network}
\seealso{
  \code{\link[igraph]{igraph}}
  \code{\link[bipartite]{bipartite}}
}
