\name{virtualComCompetition}
\docType{data}
\alias{virtualComCompetition}
\title{Dataset simulated with communities whose composition is driven
  simultaneously by biotic (competition) and abiotic environmental effects.}
\description{
  50 species in interaction in a metaweb, with presence/absence data for
  400 locations and a single enviromental variable.
}
\usage{data(virtualComCompetition)}
\format{A list containing three objects presence, metaweb and environmentvar.}
