\name{divnet-package}
\alias{divnet-package}
\alias{divnet}
\docType{package}
\title{
Diversity indices for networks
}
\description{
Bla bla bla bla bla blab  bla
}
\details{
divnet is bla bla  bla.
}
\author{
Authors: Stéphane Dray, Vincent Miele, Marc Ohlmann, Wilfried Thuiller
  
Maintainer: Wilfried Thuiller <wilfried.thuiller@univ-grenoble-alpes.fr>
}
\references{
Catherine Matias and Vincent Miele, \emph{Statistical clustering of temporal networks through a dynamic stochastic block model}, Journal of the Royal Statistical Society: Series B (2016)
http://dx.doi.org/10.1111/rssb.12200
http://arxiv.org/abs/1506.07464
}
\keyword{ package }
\keyword{ ecology }
\keyword{ diversity }
\keyword{ network }
%\seealso{
%Optional links to other man pages, e.g. \code{\link[<pkg>:<pkg>-package]{<pkg>}}
%}
%\examples{
%}
